requirements.txt
langchain
langchain_community
langchain-openai
openai
streamlit
python-dotenv
icecream
youtube-transcript-api
faiss-cpu
pytube